LOG-1771: Refactor test/runtime package into internal/runtime
