+++ internal/runtime/runtime_suite_test.go
@@ -1,0 +1,13 @@
+package runtime_test
+
+import (
+	"testing"
+
+	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/gomega"
+)
+
+func TestRuntime(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "Runtime Suite")
+}
