+++ test/client/client.go
@@ -3,12 +3,12 @@ package client
 import (
 	"context"
 	"fmt"
+	testrt "github.com/openshift/cluster-logging-operator/internal/runtime"
 	"sync"
 	"time"
 
 	"github.com/ViaQ/logerr/log"
 	"github.com/openshift/cluster-logging-operator/test"
-	testrt "github.com/openshift/cluster-logging-operator/test/runtime"
 	"k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/api/meta"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
