+++ test/client/watch.go
@@ -4,13 +4,13 @@ import (
 	"context"
 	"errors"
 	"fmt"
+	"github.com/openshift/cluster-logging-operator/internal/runtime"
 	"reflect"
 	gort "runtime"
 	"strings"
 	"time"
 
 	"github.com/ViaQ/logerr/log"
-	"github.com/openshift/cluster-logging-operator/test/runtime"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime/schema"
 	"k8s.io/apimachinery/pkg/watch"
