+++ test/functional/output_elasticsearch7.go
@@ -1,11 +1,11 @@
 package functional
 
 import (
+	"github.com/openshift/cluster-logging-operator/internal/runtime"
 	"strings"
 
 	"github.com/ViaQ/logerr/log"
 	logging "github.com/openshift/cluster-logging-operator/apis/logging/v1"
-	"github.com/openshift/cluster-logging-operator/test/runtime"
 )
 
 const (
