+++ test/functional/output_forward.go
@@ -1,13 +1,13 @@
 package functional
 
 import (
+	"github.com/openshift/cluster-logging-operator/internal/runtime"
 	"strings"
 
 	"github.com/ViaQ/logerr/log"
 	logging "github.com/openshift/cluster-logging-operator/apis/logging/v1"
 	"github.com/openshift/cluster-logging-operator/internal/constants"
 	"github.com/openshift/cluster-logging-operator/internal/utils"
-	"github.com/openshift/cluster-logging-operator/test/runtime"
 )
 
 const (
